[race] handle when more lanes than expected are given (#32)
We found a couple bugs where if more lane events than expected are
given, then the application halts. This PR fixes that problem by
filtering out spurious lane events on the client before we try to
process them.

Additionally, I show the <kbd>Redo</kbd> button as soon as we get any
lane data. This handles the case of a car running off the edge and we
need to rerun the race.